--- VarleLt/VarleLt.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VarleLt", "VarleLt\VarleLt.csproj", "{7B8F3DD6-E483-4AE6-A456-DB11320917D0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B8F3DD6-E483-4AE6-A456-DB11320917D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B8F3DD6-E483-4AE6-A456-DB11320917D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B8F3DD6-E483-4AE6-A456-DB11320917D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B8F3DD6-E483-4AE6-A456-DB11320917D0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {70380E4C-0214-4342-9B15-EDAC526F23A5}
-	EndGlobalSection
-EndGlobal
